requirements.txt
discord.py==2.3.2
psutil==5.9.8
docker==7.1.0
python-dotenv==1.0.1
matplotlib==3.8.3
